requirements.txt
streamlit==1.41.1
groq==0.11.0
python-dotenv==1.0.1
